# specifications/openapi.yaml
openapi: 3.1.0
info:
    title: PassWordManagerAPI
    version: 1.4.0
    description: >-
        PassWordManager API used to manage users, personal credential databases and
        the services offered to the user
    contact:
        email: lukaarma@gmail.com
    license:
        name: MIT
        url: "https://opensource.org/licenses/MIT"
servers:
    - url: "https://dev.lukaarma.dynu.net/api"
      description: Development environment
    - url: "https://pwm.lukaarma.dynu.net/api"
      description: Production environment
paths:
    /user/login:
        post:
            summary: Login to existing user
            security: []

            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                email:
                                    type: string
                                    format: email
                                masterPwdHash:
                                    type: string
                                    format: password
                            required:
                                - email
                                - masterPwdHash

            responses:
                "200":
                    description: Login successful
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    firstName:
                                        type: string
                                    lastName:
                                        type: string
                                    email:
                                        type: string
                                        format: email
                                    PSK:
                                        type: string
                                        format: hex
                                    IV:
                                        type: string
                                        format: hex
                                required:
                                    - firstName
                                    - PSK
                                    - IV
                "400":
                    $ref: "#/components/responses/400"
                "500":
                    $ref: "#/components/responses/500"
    /user/signup:
        post:
            summary: Register new user
            security: []

            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                email:
                                    type: string
                                    format: email
                                masterPwdHash:
                                    type: string
                                    format: hex
                                firstName:
                                    type: string
                                lastName:
                                    type: string
                                PSK:
                                    type: string
                                    format: hex
                                IV:
                                    type: string
                                    format: hex
                            required:
                                - email
                                - masterPwdHash
                                - firstName
                                - lastName
                                - PSK
                                - IV
            responses:
                "200":
                    $ref: "#/components/responses/200"
                "400":
                    $ref: "#/components/responses/400"
                "500":
                    $ref: "#/components/responses/500"
    /user/delete:
        post:
            summary: Delete the user's account. Cannot be undone!

            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                masterPwdHash:
                                    type: string
                                    format: password
                            required:
                                - masterPwdHash

            responses:
                "200":
                    $ref: "#/components/responses/200"
                "400":
                    $ref: "#/components/responses/400"
                "500":
                    $ref: "#/components/responses/500"
    /user/profile:
        get:
            summary: Get user information

            responses:
                "200":
                    $ref: "#/components/responses/UserInfoResponse"
                "500":
                    $ref: "#/components/responses/500"

        put:
            summary: Update user First and Last name

            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                firstName:
                                    type: string
                                lastName:
                                    type: string

            responses:
                "200":
                    $ref: "#/components/responses/UserInfoResponse"
                "400":
                    $ref: "#/components/responses/400"
                "500":
                    $ref: "#/components/responses/500"
    /user/sendVerification:
        post:
            summary: Send a new token if the email belongs to a user awaiting verification
            security: []

            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                email:
                                    type: string
                                    format: email
                            required:
                                - email
            responses:
                "200":
                    $ref: "#/components/responses/200"
                "400":
                    $ref: "#/components/responses/400"
                "500":
                    $ref: "#/components/responses/500"
    /user/verify/{token}:
        get:
            summary: Verify the user email and activate the account
            security: []

            parameters:
                - name: token
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
                  description: Verification token (UUIDv4) sent via mail to the user

            responses:
                "200":
                    $ref: "#/components/responses/200"
                "400":
                    $ref: "#/components/responses/400"
                "500":
                    $ref: "#/components/responses/500"
...
components:
    securitySchemes:
        JWTAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: |-
                Example:
                > Authorization: Bearer
    schemas:
        Success:
            type: object
            properties:
                code:
                    type: integer
                    min: 100
                    max: 299
                message:
                    type: string
            required:
                - code
                - message
        Error:
            type: object
            properties:
                code:
                    type: integer
                    min: 300
                    max: 1000
                message:
                    type: string
            required:
                - code
                - message
        UserInfo:
            type: object
            properties:
                email:
                    type: string
                firstName:
                    type: string
                lastName:
                    type: string
            required:
                - email
                - firstName
                - lastName
...
    responses:
        UserInfoResponse:
            description: Complete user info
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/UserInfo"
...
        "200":
            description: Request successful
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/Success"
        "400":
            description: Bad request
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/Error"
        "500":
            description: Server error
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/Error"
security:
    - JWTAuth: []